+++ src/components/datePicker.js
@@ -1,0 +1,3162 @@
+/**
+ * datePicker
+ * 日期时间选择器
+ */
+
+import { lay } from '../core/lay.js';
+import { i18n } from '../core/i18n.js';
+import { $ } from 'jquery';
+import { Popup, popupHooks } from './popup.js';
+
+export class DatePicker extends Popup {
+  static componentName = 'datePicker';
+
+  /**
+   * 默认配置项
+   * 可继承 {@link Popup.options}
+   */
+  static options = {
+    ...super.options,
+
+    // popup 组件相关选项
+    placement: 'bottom-start', // 弹出方位
+    repeatClick: 'keep', // 再次点击目标元素时保持层的当前状态
+
+    // 控件类型；可选值: year|month|date|time|datetime
+    type: 'date',
+
+    // 是否开启范围选择，即双控件
+    range: false,
+
+    // 默认日期格式
+    format: 'yyyy-MM-dd',
+
+    // 默认日期。支持传入 `new Date()` 或符合 `format` 设定的日期格式字符
+    value: null,
+
+    // 是否自动给目标元素填充初始值；需设置 `value` 后有效
+    autoFillValue: true,
+
+    // 一周的开始是周几。0 表示周日，1 表示周一，以此类推
+    weekStart: 0,
+
+    // 有效最小日期，年月日必须用「-」分割，时分秒必须用「:」分割
+    // 注：它并不遵循 format 设定的格式
+    min: '1900-1-1',
+
+    // 有效最大日期，同上
+    max: '2099-12-31',
+
+    // 是否开启完整面板；开启后，将同时显示快捷栏、日期栏、时间栏
+    // 仅当 `type: 'datetime'` 且未开启 `range` 属性时生效
+    fullPanel: false,
+
+    // 是否显示底部栏
+    showBottom: true,
+
+    // 是否显示值预览
+    showPreview: true,
+
+    // 右下角显示的按钮，会按照数组顺序排列
+    btns: ['clear', 'now', 'confirm'],
+
+    // 是否自动确认；仅在未开启 `range` 时有效
+    autoConfirm: true,
+
+    // 是否将面板内嵌渲染至目标容器中
+    inline: false,
+
+    // 语言，只内置支持中英文( cn|en )，其他语言可通过 i18n 扩展
+    // 为实现 lang 选项就近生效，去除此处的默认值，$t 设置了英文回退值
+    lang: '',
+
+    // 样式变体。可选值: default|solid|grid
+    variant: 'default',
+
+    // 是否显示公历重要节日，仅支持中文版
+    showFestival: false,
+
+    // 日期备注，如重要事件或活动标记
+    mark: {},
+
+    // 标注法定节假日或补假上班
+    // holidays: null,
+
+    // 控件选择完毕后的回调，点击清空/现在/确定也均会触发
+    // done: null,
+
+    // 日期时间改变后的回调
+    // onChange: null,
+  };
+
+  static get CONST() {
+    return {
+      ...super.CONST,
+      ELEM: 'lay-datepicker',
+      ELEM_STATIC: 'lay-datepicker-static',
+      ELEM_LIST: 'lay-datepicker-list',
+      ELEM_SELECTED: 'lay-datepicker-selected',
+      ELEM_HINT: 'lay-datepicker-hint',
+      ELEM_DAY_NOW: 'lay-datepicker-day-now',
+      ELEM_PREV: 'lay-datepicker-day-prev',
+      ELEM_NEXT: 'lay-datepicker-day-next',
+      ELEM_FOOTER: 'lay-datepicker-footer',
+      ELEM_SHORTCUT: 'lay-datepicker-shortcut',
+      ELEM_NOW: 'lay-datepicker-btns-now',
+      ELEM_CONFIRM: 'lay-datepicker-btns-confirm',
+      ELEM_TIME_TEXT: 'lay-datepicker-time-text',
+      ELEM_TIME_BTN: 'lay-datepicker-btns-time',
+      ELEM_PREVIEW: 'lay-datepicker-preview',
+      ELEM_MAIN: 'lay-datepicker-main',
+      LIMIT_YEAR: [100, 200000],
+    };
+  }
+
+  // 实例方法静态委托
+  static {
+    this.delegateInstanceMethods(['hint', 'unbind']);
+  }
+
+  /**
+   * 将组件内部日期对象按模板格式化为字符串
+   * @param {Object} dateTime - 组件内部日期对象
+   * @param {string|Array} format - 日期格式模板或已解析的格式数组
+   * @param {boolean} monthOffset - 月份偏移量（传 1 表示月份 +1）
+   * @returns {string} - 格式化后的日期字符串
+   */
+  static formatDateTime(dateTime = {}, format, monthOffset) {
+    // 如果 format 是字符型，则转换为数组格式
+    if (typeof format === 'string') {
+      format = splitDateFormat(format);
+    }
+
+    format = (format || []).concat();
+
+    // 转义为规定格式
+    format.forEach(function (item, i) {
+      if (/yyyy|y/.test(item)) {
+        // 年
+        format[i] = lay.digit(dateTime.year, item.length);
+      } else if (/MM|M/.test(item)) {
+        // 月
+        format[i] = lay.digit(dateTime.month + (monthOffset || 0), item.length);
+      } else if (/dd|d/.test(item)) {
+        // 日
+        format[i] = lay.digit(dateTime.date, item.length);
+      } else if (/HH|H/.test(item)) {
+        // 时
+        format[i] = lay.digit(dateTime.hours, item.length);
+      } else if (/mm|m/.test(item)) {
+        // 分
+        format[i] = lay.digit(dateTime.minutes, item.length);
+      } else if (/ss|s/.test(item)) {
+        // 秒
+        format[i] = lay.digit(dateTime.seconds, item.length);
+      }
+    });
+
+    return format.join('');
+  }
+
+  /**
+   * 获取指定月份的天数
+   * @param {number} month - 月份（0-11）
+   * @param {number} year - 年份
+   * @returns {number} - 指定月份的天数
+   */
+  static getDaysInMonth(month, year) {
+    const date = new Date();
+
+    // 设置日期为下个月的第一天
+    date.setFullYear(
+      year || date.getFullYear(),
+      month || date.getMonth() + 1,
+      1,
+    );
+
+    // 减去一天，得到当前月最后一天
+    return new Date(date.getTime() - 1000 * 60 * 60 * 24).getDate();
+  }
+
+  // 构造函数
+  constructor(options) {
+    super(options);
+
+    this.overrideArrayOptions(options);
+
+    // 更新 i18n 消息对象
+    this.i18nMessages = this.getI18nMessages();
+
+    // 处理日期面板顶部年月顺序
+    // 这是一个变通的方法，因为 i18nMessages.monthBeforeYear 不存在
+    if (typeof this.i18nMessages.monthBeforeYear !== 'boolean') {
+      if (!window.Intl) {
+        this.i18nMessages.monthBeforeYear = !(
+          YearBeforeMonthLocale.indexOf(options.lang) > -1
+        );
+      } else {
+        var formatter = new Intl.DateTimeFormat(options.lang, {
+          year: 'numeric',
+          month: 'short',
+        });
+        var parts = formatter.formatToParts(new Date(1970, 0));
+        var order = [];
+        parts.map(function (part) {
+          if (part.type === 'year' || part.type === 'month') {
+            order.push(part.type);
+          }
+        });
+        this.i18nMessages.monthBeforeYear = order[0] === 'month';
+      }
+    }
+  }
+
+  /**
+   * 层渲染时的内部钩子
+   * @returns {void}
+   */
+  [popupHooks.kOnRender]() {
+    const { options } = this;
+
+    this.#prepare();
+
+    // 内嵌模式
+    if (options.inline) {
+      const mainElem = this.#buildMainPanel();
+      this.$mainElem.addClass('lay-panel lay-border-box');
+      options.$elem.html(mainElem);
+    }
+
+    // 默认赋值
+    if (options.value && options.autoFillValue) {
+      let value = options.value;
+      if (lay.type(value) === 'date') {
+        value = this.parse(0, this.systemDate(options.value));
+      }
+      this.setValue(value);
+    }
+  }
+
+  /**
+   * 层打开后的内部钩子
+   * @returns {void}
+   */
+  [popupHooks.kAfterOpen]() {
+    const { options, constructor } = this;
+
+    if (options.inline) return;
+
+    const $document = $(document);
+    const eventNamespace = constructor.CONST.EVENT_NAMESPACE;
+
+    // 回车触发确认
+    $document
+      .off(`keydown${eventNamespace}`)
+      .on(`keydown${eventNamespace}`, (e) => {
+        if (e.keyCode !== 13 || !this.isRootElemMounted()) return;
+        e.preventDefault();
+        const confirmBtn = $(this.footer).find(`.${CONST.ELEM_CONFIRM}`)[0];
+        if (confirmBtn) {
+          confirmBtn.click();
+        }
+      });
+  }
+
+  /**
+   * 层打开前的内部钩子
+   * @returns {void}
+   */
+  [popupHooks.kBeforeOpen]() {
+    const { options, $rootElem } = this;
+
+    if (options.inline) return;
+
+    const mainElem = this.#buildMainPanel();
+
+    // 设置 Popup 层内容
+    options.content = mainElem;
+
+    // 添加组件专属 className
+    $rootElem.addClass(CONST.ELEM);
+  }
+
+  // 层关闭后的内部钩子
+  [popupHooks.kAfterClose]() {
+    const { $mainElem, constructor } = this;
+    const eventNamespace = constructor.CONST.EVENT_NAMESPACE;
+
+    if (!$mainElem.hasClass(CONST.ELEM_STATIC)) {
+      this.checkDate();
+      delete this.startDate;
+      delete this.endDate;
+      delete this.endState;
+      delete this.startTime;
+      delete this.endTime;
+    }
+
+    $(document).off(`keydown${eventNamespace}`);
+  }
+
+  /**
+   * 规范化选项
+   * @returns {void}
+   */
+  #normalizeOptions() {
+    const options = this.options;
+
+    // 若开启完整面板
+    if (options.fullPanel) {
+      if (options.type !== 'datetime' || options.range) {
+        // 目前只支持 datetime 的全面版
+        options.fullPanel = false;
+      }
+    }
+
+    // 若 type 非法，则重置为默认值
+    if (!defaultFormat[options.type]) {
+      console.error(
+        `datePicker type error: '${options.type}' is not supported`,
+      );
+      options.type = this.constructor.options.type;
+    }
+
+    // 根据不同 type，初始化默认 format
+    if (options.format === defaultFormat.date) {
+      options.format = defaultFormat[options.type] || defaultFormat.date;
+    }
+
+    // 设置了一周的开始是周几，此处做一个控制
+    if (options.weekStart) {
+      if (!/^[0-6]$/.test(options.weekStart)) {
+        var lang = this.i18nMessages;
+        options.weekStart = lang.weeks.indexOf(options.weekStart);
+        if (options.weekStart === -1) options.weekStart = 0;
+      }
+    }
+
+    // 如果不是 input|textarea 元素，则采用默认 click 事件
+    if (!this.isInput()) {
+      if (options.trigger === 'focus') {
+        options.trigger = 'click';
+      }
+    }
+
+    // 获取限制内日期
+    ['min', 'max'].forEach((item, i) => {
+      var ymd = [];
+      var hms = [];
+      if (typeof options[item] === 'number') {
+        // 如果为数字
+        var day = options[item],
+          tDate = new Date(),
+          time = this.newDate({
+            // 今天的最大毫秒数
+            year: tDate.getFullYear(),
+            month: tDate.getMonth(),
+            date: tDate.getDate(),
+            hours: i ? 23 : 0,
+            minutes: i ? 59 : 0,
+            seconds: i ? 59 : 0,
+          }).getTime(),
+          STAMP = 86400000, // 代表一天的毫秒数
+          thisDate = new Date(
+            day
+              ? day < STAMP
+                ? time + day * STAMP
+                : day // 如果数字小于一天的毫秒数，则数字为天数，否则为毫秒数
+              : time,
+          );
+        ymd = [
+          thisDate.getFullYear(),
+          thisDate.getMonth() + 1,
+          thisDate.getDate(),
+        ];
+        hms = [
+          thisDate.getHours(),
+          thisDate.getMinutes(),
+          thisDate.getSeconds(),
+        ];
+      } else if (typeof options[item] === 'string') {
+        ymd = (options[item].match(/\d+-\d+-\d+/) || [''])[0].split('-');
+        hms = (options[item].match(/\d+:\d+:\d+/) || [''])[0].split(':');
+      } else if (typeof options[item] === 'object') {
+        return options[item];
+      }
+      options[item] = {
+        year: ymd[0] | 0 || new Date().getFullYear(),
+        month: ymd[1] ? (ymd[1] | 0) - 1 : new Date().getMonth(),
+        date: ymd[2] | 0 || new Date().getDate(),
+        hours: hms[0] | 0,
+        minutes: hms[1] | 0,
+        seconds: hms[2] | 0,
+      };
+    });
+  }
+
+  /**
+   * 组件预处理
+   * @returns {void}
+   */
+  #prepare() {
+    var that = this;
+    const options = this.options;
+
+    this.#normalizeOptions();
+
+    // 日期范围分隔符
+    that.rangeStr = options.range
+      ? typeof options.range === 'string'
+        ? options.range
+        : '-'
+      : '';
+
+    // 日期范围的日历面板是否联动
+    that.rangeLinked = !!(
+      options.range &&
+      options.rangeLinked &&
+      (options.type === 'date' || options.type === 'datetime')
+    );
+
+    // 切换日历联动方式
+    that.autoCalendarModel = function () {
+      var state = that.rangeLinked;
+      that.rangeLinked =
+        options.range &&
+        (options.type === 'date' || options.type === 'datetime') &&
+        (!that.startDate ||
+          !that.endDate ||
+          (that.startDate &&
+            that.endDate &&
+            that.startDate.year === that.endDate.year &&
+            that.startDate.month === that.endDate.month));
+      that.$mainElem[that.rangeLinked ? 'addClass' : 'removeClass'](
+        'lay-datepicker-linkage',
+      );
+      return that.rangeLinked != state; // 返回发生了变化
+    };
+
+    // 是否自动切换
+    that.autoCalendarModel.auto =
+      that.rangeLinked && options.rangeLinked === 'auto';
+
+    // 若 range 参数为数组，则表示为开始日期和结束日期的 input 对象
+    if (lay.type(options.range) === 'array') {
+      that.rangeElem = [$(options.range[0]), $(options.range[1])];
+    }
+
+    // 将日期格式模板拆分成数组
+    that.format = splitDateFormat(options.format);
+
+    // 生成正则表达式
+    that.EXP_IF = '';
+    that.EXP_SPLIT = '';
+    that.format.forEach(function (item, i) {
+      var EXP = new RegExp(DATE_FORMAT_REGEX).test(item)
+        ? '\\d{' +
+          (function () {
+            if (
+              new RegExp(DATE_FORMAT_REGEX).test(
+                that.format[i === 0 ? i + 1 : i - 1] || '',
+              )
+            ) {
+              if (/^yyyy|y$/.test(item)) return 4;
+              return item.length;
+            }
+            if (/^yyyy$/.test(item)) return '1,4';
+            if (/^y$/.test(item)) return '1,308';
+            return '1,2';
+          })() +
+          '}'
+        : '\\' + item;
+      that.EXP_IF = that.EXP_IF + EXP;
+      that.EXP_SPLIT = that.EXP_SPLIT + '(' + EXP + ')';
+    });
+    // 验证日期格式正则
+    that.EXP_IF_ONE = new RegExp('^' + that.EXP_IF + '$'); // 验证单个日期格式
+    that.EXP_IF = new RegExp(
+      '^' +
+        (options.range
+          ? that.EXP_IF + '\\s\\' + that.rangeStr + '\\s' + that.EXP_IF
+          : that.EXP_IF) +
+        '$',
+    );
+    that.EXP_SPLIT = new RegExp('^' + that.EXP_SPLIT + '$', '');
+
+    // 重定义 input 元素的 get set
+    if (typeof options.formatToDisplay === 'function') {
+      if (that.isInput()) {
+        that.formatToDisplay(options.$elem[0], options.formatToDisplay);
+      } else {
+        var rangeElem = that.rangeElem;
+        if (rangeElem) {
+          that.formatToDisplay(rangeElem[0][0], options.formatToDisplay);
+          that.formatToDisplay(rangeElem[1][0], options.formatToDisplay);
+        }
+      }
+    }
+  }
+
+  getI18nMessages() {
+    var that = this;
+    var options = that.options;
+    var locale = i18n.config.locale;
+
+    // 纠正旧版「简体中文」语言码
+    if (options.lang === 'cn') {
+      options.lang = zhCN;
+    } else if (!options.lang) {
+      options.lang = i18n.config.locale;
+    }
+    locale = options.lang;
+
+    return {
+      months: i18n.$t('datePicker.months', null, {
+        locale: locale,
+        default: [
+          'Jan',
+          'Feb',
+          'Mar',
+          'Apr',
+          'May',
+          'Jun',
+          'Jul',
+          'Aug',
+          'Sep',
+          'Oct',
+          'Nov',
+          'Dec',
+        ],
+      }),
+      weeks: i18n.$t('datePicker.weeks', null, {
+        locale: locale,
+        default: ['Su', 'Mo', 'Tu', 'We', 'Th', 'Fr', 'Sa'],
+      }),
+      time: i18n.$t('datePicker.time', null, {
+        locale: locale,
+        default: ['Hour', 'Minute', 'Second'],
+      }),
+      literal: {
+        year: i18n.$t('datePicker.literal.year', null, {
+          locale: locale,
+          default: '',
+        }),
+      },
+      monthBeforeYear: i18n.$t('datePicker.monthBeforeYear', null, {
+        locale: locale,
+        default: null,
+      }),
+      selectDate: i18n.$t('datePicker.selectDate', null, {
+        locale: locale,
+        default: 'Select Date',
+      }),
+      selectTime: i18n.$t('datePicker.selectTime', null, {
+        locale: locale,
+        default: 'Select Time',
+      }),
+      startTime: i18n.$t('datePicker.startTime', null, {
+        locale: locale,
+        default: 'Start Time',
+      }),
+      endTime: i18n.$t('datePicker.endTime', null, {
+        locale: locale,
+        default: 'End Time',
+      }),
+      tools: {
+        confirm: i18n.$t('datePicker.tools.confirm', null, {
+          locale: locale,
+          default: 'Confirm',
+        }),
+        clear: i18n.$t('datePicker.tools.clear', null, {
+          locale: locale,
+          default: 'Clear',
+        }),
+        now: i18n.$t('datePicker.tools.now', null, {
+          locale: locale,
+          default: 'Now',
+        }),
+        reset: i18n.$t('datePicker.tools.reset', null, {
+          locale: locale,
+          default: 'Reset',
+        }),
+      },
+      rangeOrderPrompt: i18n.$t('datePicker.rangeOrderPrompt', null, {
+        locale: locale,
+        default: 'End time cannot be less than start Time\nPlease re-select',
+      }),
+      invalidDatePrompt: i18n.$t('datePicker.invalidDatePrompt', null, {
+        locale: locale,
+        default: 'Invalid date\n',
+      }),
+      formatErrorPrompt: function (format) {
+        return i18n.$t(
+          'datePicker.formatErrorPrompt',
+          { format: format },
+          {
+            locale: locale,
+            default:
+              'Date format is invalid\nMust follow the format:\n{format}\n',
+          },
+        );
+      },
+      autoResetPrompt: i18n.$t('datePicker.autoResetPrompt', null, {
+        locale: locale,
+        default: 'It has been reset',
+      }),
+      preview: i18n.$t('datePicker.preview', null, {
+        locale: locale,
+        default: 'The selected result',
+      }),
+    };
+  }
+
+  // 仅简体中文生效，不做国际化
+  markerOfChineseFestivals = {
+    '0-1-1': '元旦',
+    '0-3-8': '妇女',
+    '0-3-12': '植树',
+    '0-4-1': '愚人',
+    '0-5-1': '劳动',
+    '0-5-4': '青年',
+    '0-6-1': '儿童',
+    '0-9-10': '教师',
+    '0-10-1': '国庆',
+    '0-12-25': '圣诞',
+  };
+
+  /**
+   * 构建主面板
+   * @returns {HTMLElement} 主面板元素
+   */
+  #buildMainPanel() {
+    var that = this;
+    var options = that.options;
+    var lang = that.i18nMessages;
+
+    var mainElem = lay.elem('div', {
+      [CONST.ATTR_ID]: options.id,
+      class: (() => {
+        const classLists = [CONST.ELEM];
+        if (options.range) classLists.push('lay-datepicker-range');
+        if (that.rangeLinked) classLists.push('lay-datepicker-linkage');
+        if (options.inline) classLists.push(CONST.ELEM_STATIC);
+        if (options.fullPanel)
+          classLists.push('lay-datepicker-variant-fullpanel');
+        if (options.variant) {
+          classLists.push('lay-datepicker-variant-' + options.variant);
+        }
+        return classLists.join(' ');
+      })(),
+    });
+
+    // 主体区域
+    var mainListElems = (that.mainListElems = []);
+    var headerElems = (that.headerElems = []);
+    var contentElems = (that.contentElems = []);
+    var tableElems = (that.tableElems = []);
+
+    // 底部区域
+    var divFooter = (that.footer = lay.elem('div', {
+      class: CONST.ELEM_FOOTER,
+    }));
+
+    // 快捷栏区域
+    var divShortcut = (that.shortcut = lay.elem('ul', {
+      class: CONST.ELEM_SHORTCUT,
+    }));
+
+    that.$mainElem = $(mainElem);
+
+    // 单双日历区域
+    for (var i = 0; i < 2; i++) {
+      if (!options.range && i > 0) {
+        break;
+      }
+
+      // 头部区域
+      var divHeader = lay.elem('div', {
+          class: 'lay-datepicker-header',
+        }),
+        // 左右切换
+        headerChild = [
+          (function () {
+            // 上一年
+            var elem = lay.elem('i', {
+              class: 'lay-icon lay-datepicker-icon lay-datepicker-prev-y',
+            });
+            elem.innerHTML = '&#xe65a;';
+            return elem;
+          })(),
+          (function () {
+            // 上一月
+            var elem = lay.elem('i', {
+              class: 'lay-icon lay-datepicker-icon lay-datepicker-prev-m',
+            });
+            elem.innerHTML = '&#xe603;';
+            return elem;
+          })(),
+          (function () {
+            // 年月选择
+            var elem = lay.elem('div', {
+              class: 'lay-datepicker-set-ym',
+            });
+            elem.appendChild(lay.elem('span'));
+            elem.appendChild(lay.elem('span'));
+            return elem;
+          })(),
+          (function () {
+            // 下一月
+            var elem = lay.elem('i', {
+              class: 'lay-icon lay-datepicker-icon lay-datepicker-next-m',
+            });
+            elem.innerHTML = '&#xe602;';
+            return elem;
+          })(),
+          (function () {
+            // 下一年
+            var elem = lay.elem('i', {
+              class: 'lay-icon lay-datepicker-icon lay-datepicker-next-y',
+            });
+            elem.innerHTML = '&#xe65b;';
+            return elem;
+          })(),
+        ],
+        // 日历内容区域
+        divContent = lay.elem('div', {
+          class: 'lay-datepicker-content',
+        }),
+        table = lay.elem('table'),
+        thead = lay.elem('thead'),
+        theadTr = lay.elem('tr');
+
+      // 生成年月选择
+      headerChild.forEach(function (item) {
+        divHeader.appendChild(item);
+      });
+
+      // 生成表格
+      thead.appendChild(theadTr);
+      for (var rowIndex = 0; rowIndex < 6; rowIndex++) {
+        // 表体
+        var tr = table.insertRow(0);
+        for (var j = 0; j < 7; j++) {
+          if (rowIndex === 0) {
+            var th = lay.elem('th');
+            th.innerHTML = lang.weeks[(j + options.weekStart) % 7];
+            theadTr.appendChild(th);
+          }
+          tr.insertCell(j);
+        }
+      }
+      table.insertBefore(thead, table.children[0]); // 表头
+      divContent.appendChild(table);
+
+      mainListElems[i] = lay.elem('div', {
+        class: CONST.ELEM_MAIN + ' lay-datepicker-main-list-' + i,
+      });
+
+      mainListElems[i].appendChild(divHeader);
+      mainListElems[i].appendChild(divContent);
+
+      headerElems.push(headerChild);
+      contentElems.push(divContent);
+      tableElems.push(table);
+    }
+
+    // 生成底部栏
+    $(divFooter).html(
+      (function () {
+        var html = [],
+          btns = [];
+        if (options.type === 'datetime') {
+          html.push(
+            '<span lay-type="datetime" class="' +
+              CONST.ELEM_TIME_BTN +
+              '">' +
+              lang.selectTime +
+              '</span>',
+          );
+        }
+        if (
+          !(!options.range && options.type === 'datetime') ||
+          options.fullPanel
+        ) {
+          html.push(
+            '<span class="' +
+              CONST.ELEM_PREVIEW +
+              '" title="' +
+              lang.preview +
+              '"></span>',
+          );
+        }
+
+        options.btns.forEach(function (item) {
+          var title = lang.tools[item] || 'btn';
+          if (options.range && item === 'now') return;
+          if (options.inline && item === 'clear') {
+            title = lang.tools.reset;
+          }
+          btns.push(
+            '<span lay-type="' +
+              item +
+              '" class="lay-datepicker-btns-' +
+              item +
+              '">' +
+              title +
+              '</span>',
+          );
+        });
+        html.push(
+          '<div class="lay-datepicker-footer-btns">' + btns.join('') + '</div>',
+        );
+        return html.join('');
+      })(),
+    );
+
+    // 生成快捷键栏
+    if (options.shortcuts) {
+      mainElem.appendChild(divShortcut);
+      $(divShortcut)
+        .html(
+          (function () {
+            var shortcutBtns = [];
+            options.shortcuts.forEach(function (item, i) {
+              shortcutBtns.push(
+                '<li data-index="' + i + '">' + item.text + '</li>',
+              );
+            });
+            return shortcutBtns.join('');
+          })(),
+        )
+        .find('li')
+        .on('click', function () {
+          var btnSetting = options.shortcuts[this.dataset['index']] || {};
+          var value =
+            (typeof btnSetting.value === 'function'
+              ? btnSetting.value()
+              : btnSetting.value) || [];
+          if (!lay.isArray(value)) {
+            value = [value];
+          }
+          var type = options.type;
+          value.forEach(function (item, i) {
+            var dateTime = [options.dateTime, that.endDate][i];
+            if (type === 'time' && lay.type(item) !== 'date') {
+              if (that.EXP_IF.test(item)) {
+                item = (item.match(that.EXP_SPLIT) || []).slice(1);
+                lay.extend(dateTime, {
+                  hours: item[0] | 0,
+                  minutes: item[2] | 0,
+                  seconds: item[4] | 0,
+                });
+              }
+            } else {
+              lay.extend(
+                dateTime,
+                that.systemDate(
+                  lay.type(item) === 'date' ? item : new Date(item),
+                ),
+              );
+            }
+
+            if (type === 'time' || type === 'datetime') {
+              that[['startTime', 'endTime'][i]] = {
+                hours: dateTime.hours,
+                minutes: dateTime.minutes,
+                seconds: dateTime.seconds,
+              };
+            }
+            if (i === 0) {
+              // 第一个值作为 startDate
+              that.startDate = lay.extend({}, dateTime);
+            } else {
+              that.endState = true;
+            }
+            if (type === 'year' || type === 'month' || type === 'time') {
+              that.listYM[i] = [dateTime.year, dateTime.month + 1];
+            } else if (i) {
+              that.autoCalendarModel.auto && that.autoCalendarModel();
+            }
+          });
+          that.checkDate('limit').calendar(null, null, 'init');
+
+          var timeBtn = $(that.footer)
+            .find('.' + CONST.ELEM_TIME_BTN)
+            .removeClass(CONST.CLASS_IS_DISABLED);
+          timeBtn && timeBtn.attr('lay-type') === 'date' && timeBtn[0].click();
+          that.done('onChange');
+
+          $(this).addClass(CONST.CLASS_IS_SELECTED);
+
+          // 自动确认
+          if (!options.inline) {
+            that.setValue(that.parse()).done();
+            that.close();
+          }
+          /*
+          if (!options.inline && !options.range && options.autoConfirm) {
+            if (type === 'date') {
+              that.choose($(elem).find('td.lay-this'))
+            } else if (type === 'year' || type === 'month') {
+              if($(mainListElems[0]).find('.' + CONST.ELEM_MAIN + ' li.' + CONST.CLASS_IS_SELECTED + ':not(.lay-datepicker-disabled)')[0]) {
+                that.setValue(that.parse()).done();
+                that.close();
+              }
+            }
+          }
+          */
+        });
+    }
+
+    // 插入到主区域
+    mainListElems.forEach(function (main) {
+      mainElem.appendChild(main);
+    });
+    options.showBottom && mainElem.appendChild(divFooter);
+
+    this.checkDate().calendar(null, 0, 'init');
+    that.changeEvent(); // 日期切换
+
+    that.renderAdditional();
+    typeof options.ready === 'function' &&
+      options.ready(
+        lay.extend({}, options.dateTime, {
+          month: options.dateTime.month + 1,
+        }),
+      );
+
+    that.preview();
+
+    return mainElem;
+  }
+
+  // 提示
+  hint(opts = {}) {
+    var that = this;
+    var div = lay.elem('div', {
+      class: CONST.ELEM_HINT,
+    });
+    const $mainElem = that.$mainElem;
+
+    opts = {
+      content: '',
+      ms: 3000,
+      ...opts,
+    };
+
+    if (!$mainElem) return;
+
+    div.innerHTML = opts.content;
+    $mainElem.find('.' + CONST.ELEM_HINT).remove();
+    $mainElem.append(div);
+
+    clearTimeout(that.#hintTimer);
+    that.#hintTimer = setTimeout(() => {
+      $mainElem.find('.' + CONST.ELEM_HINT).remove();
+    }, opts.ms);
+  }
+
+  #hintTimer = null;
+
+  // 获取递增/减后的年月
+  getAsYM(Y, M, type) {
+    type ? M-- : M++;
+    if (M < 0) {
+      M = 11;
+      Y--;
+    }
+    if (M > 11) {
+      M = 0;
+      Y++;
+    }
+    return [Y, M];
+  }
+
+  // 系统日期
+  systemDate(newDate) {
+    var thisDate = newDate || new Date();
+    return {
+      year: thisDate.getFullYear(), // 年
+      month: thisDate.getMonth(), // 月
+      date: thisDate.getDate(), // 日
+      hours: newDate ? newDate.getHours() : 0, // 时
+      minutes: newDate ? newDate.getMinutes() : 0, // 分
+      seconds: newDate ? newDate.getSeconds() : 0, // 秒
+    };
+  }
+
+  // 日期校验
+  checkDate(type) {
+    var that = this;
+    const Constructor = that.constructor;
+    var options = that.options,
+      lang = that.i18nMessages,
+      dateTime = (options.dateTime = options.dateTime || that.systemDate()),
+      thisMaxDate,
+      error,
+      elem = options.$elem[0],
+      // valType = that.isInput(elem) ? 'val' : 'html',
+      value = (function () {
+        // 如果传入了开始和结束日期的 input 对象，则将其拼接为日期范围字符
+        if (that.rangeElem) {
+          var vals = [that.rangeElem[0].val(), that.rangeElem[1].val()];
+
+          if (vals[0] && vals[1]) {
+            return vals.join(' ' + that.rangeStr + ' ');
+          }
+        }
+        return that.isInput(elem)
+          ? elem.value
+          : options.inline
+            ? ''
+            : $(elem).attr('lay-datepicker');
+      })(),
+      // 校验日期有效数字
+      checkValid = function (dateTime) {
+        if (!dateTime) {
+          return;
+        }
+        if (dateTime.year > CONST.LIMIT_YEAR[1])
+          ((dateTime.year = CONST.LIMIT_YEAR[1]), (error = true)); // 不能超过 20 万年
+        if (dateTime.month > 11) ((dateTime.month = 11), (error = true));
+        if (dateTime.seconds > 59)
+          ((dateTime.seconds = 0), dateTime.minutes++, (error = true));
+        if (dateTime.minutes > 59)
+          ((dateTime.minutes = 0), dateTime.hours++, (error = true));
+        if (dateTime.hours > 23) ((dateTime.hours = 0), (error = true));
+
+        // 计算当前月的最后一天
+        thisMaxDate = Constructor.getDaysInMonth(
+          dateTime.month + 1,
+          dateTime.year,
+        );
+        if (dateTime.date > thisMaxDate)
+          ((dateTime.date = thisMaxDate), (error = true));
+      },
+      // 获得初始化日期值
+      initDate = function (dateTime, value, index) {
+        var startEnd = ['startTime', 'endTime'];
+        value = (value.match(that.EXP_SPLIT) || []).slice(1);
+        index = index || 0;
+
+        if (options.range) {
+          that[startEnd[index]] = that[startEnd[index]] || {};
+        }
+        that.format.forEach(function (item, i) {
+          var thisv = parseFloat(value[i]);
+          if (value[i].length < item.length) error = true;
+          if (/yyyy|y/.test(item)) {
+            // 年
+            if (thisv < CONST.LIMIT_YEAR[0])
+              ((thisv = CONST.LIMIT_YEAR[0]), (error = true)); // 年不能低于 100 年
+            dateTime.year = thisv;
+          } else if (/MM|M/.test(item)) {
+            // 月
+            if (thisv < 1) ((thisv = 1), (error = true));
+            dateTime.month = thisv - 1;
+          } else if (/dd|d/.test(item)) {
+            // 日
+            if (thisv < 1) ((thisv = 1), (error = true));
+            dateTime.date = thisv;
+          } else if (/HH|H/.test(item)) {
+            // 时
+            if (thisv < 0) ((thisv = 0), (error = true));
+            if (thisv > 23) ((thisv = 23), (error = true));
+            dateTime.hours = thisv;
+            options.range && (that[startEnd[index]].hours = thisv);
+          } else if (/mm|m/.test(item)) {
+            // 分
+            if (thisv < 0) ((thisv = 0), (error = true));
+            if (thisv > 59) ((thisv = 59), (error = true));
+            dateTime.minutes = thisv;
+            options.range && (that[startEnd[index]].minutes = thisv);
+          } else if (/ss|s/.test(item)) {
+            // 秒
+            if (thisv < 0) ((thisv = 0), (error = true));
+            if (thisv > 59) ((thisv = 59), (error = true));
+            dateTime.seconds = thisv;
+            options.range && (that[startEnd[index]].seconds = thisv);
+          }
+        });
+        checkValid(dateTime);
+      };
+
+    if (type === 'limit') {
+      if (options.range) {
+        checkValid(that.rangeLinked ? that.startDate : dateTime); // 校验开始时间
+        that.endDate && checkValid(that.endDate); // 校验结束时间
+      } else {
+        checkValid(dateTime);
+      }
+      return that;
+    }
+
+    value = value || options.value;
+    if (typeof value === 'string') {
+      value = value.replace(/\s+/g, ' ').replace(/^\s|\s$/g, '');
+    }
+
+    // 如果开启范围，则计算结束日期
+    var getEndDate = function () {
+      if (options.range) {
+        that.endDate =
+          that.endDate ||
+          lay.extend(
+            {},
+            options.dateTime,
+            (function () {
+              var obj = {},
+                dateTime = options.dateTime,
+                EYM = that.getAsYM(dateTime.year, dateTime.month);
+
+              // 初始右侧面板的年月
+              if (options.type === 'year') {
+                obj.year = dateTime.year + 1;
+              } else if (options.type !== 'time') {
+                obj.year = EYM[0];
+                obj.month = EYM[1];
+              }
+
+              // 初始右侧面板的时间
+              if (options.type === 'datetime' || options.type === 'time') {
+                obj.hours = 23;
+                obj.minutes = obj.seconds = 59;
+              }
+
+              return obj;
+            })(),
+          );
+      }
+    };
+    getEndDate();
+
+    if (typeof value === 'string' && value) {
+      if (that.EXP_IF.test(value)) {
+        // 校验日期格式
+        if (options.range) {
+          value = value.split(' ' + that.rangeStr + ' ');
+          [options.dateTime, that.endDate].forEach(function (item, i) {
+            initDate(item, value[i], i);
+          });
+        } else {
+          initDate(dateTime, value);
+        }
+      } else {
+        // 格式不合法
+        that.hint({
+          content:
+            lang.formatErrorPrompt(
+              options.range
+                ? options.format + ' ' + that.rangeStr + ' ' + options.format
+                : options.format,
+            ) + lang.autoResetPrompt,
+        });
+        error = true;
+      }
+    } else if (value && lay.type(value) === 'date') {
+      // 若值为日期对象
+      options.dateTime = that.systemDate(value);
+    } else {
+      // 重置开始日期
+      options.dateTime = that.systemDate();
+      delete that.startTime;
+
+      // 重置结束日期
+      delete that.endDate; // 删除原有的结束日期
+      getEndDate(); // 并重新获得新的结束日期
+      delete that.endTime;
+    }
+
+    // 从日期范围表单中获取初始值
+    (function () {
+      if (that.rangeElem) {
+        var vals = [that.rangeElem[0].val(), that.rangeElem[1].val()],
+          arrDate = [options.dateTime, that.endDate];
+        vals.forEach(function (_v, _i) {
+          if (that.EXP_IF_ONE.test(_v)) {
+            // 校验日期格式
+            initDate(arrDate[_i], _v, _i);
+          }
+        });
+      }
+    })();
+
+    // 校验日期有效数字
+    checkValid(dateTime);
+    if (options.range) checkValid(that.endDate);
+
+    // 如果初始值格式错误，则纠正初始值
+    if (error && value) {
+      that.setValue(
+        options.range ? (that.endDate ? that.parse() : '') : that.parse(),
+      );
+    }
+
+    // 如果当前日期不在设定的最大小日期区间，则自动纠正在可选区域
+    // 校验主面板是否在可选日期区间
+    var minMaxError;
+    if (that.getDateTime(dateTime) > that.getDateTime(options.max)) {
+      // 若超出最大日期
+      dateTime = options.dateTime = lay.extend({}, options.max);
+      minMaxError = true;
+    } else if (that.getDateTime(dateTime) < that.getDateTime(options.min)) {
+      // 若少于最小日期
+      dateTime = options.dateTime = lay.extend({}, options.min);
+      minMaxError = true;
+    }
+
+    // 校验右侧面板是否在可选日期区间
+    if (options.range) {
+      if (
+        that.getDateTime(that.endDate) < that.getDateTime(options.min) ||
+        that.getDateTime(that.endDate) > that.getDateTime(options.max)
+      ) {
+        that.endDate = lay.extend({}, options.max);
+        minMaxError = true;
+      }
+      // 有时间范围的情况下初始化 startTime 和 endTime
+      that.startTime = {
+        hours: options.dateTime.hours,
+        minutes: options.dateTime.minutes,
+        seconds: options.dateTime.seconds,
+      };
+      that.endTime = {
+        hours: that.endDate.hours,
+        minutes: that.endDate.minutes,
+        seconds: that.endDate.seconds,
+      };
+      // 如果是年月范围，将对应的日期统一成当月的 1 日进行比较，避免出现同一个月但是开始日期大于结束日期的情况
+      if (options.type === 'month') {
+        options.dateTime.date = 1;
+        that.endDate.date = 1;
+      }
+    }
+
+    // 初始值不在最大最小范围内
+    if (minMaxError && value) {
+      that.setValue(that.parse());
+      that.hint({
+        content: `value ${lang.invalidDatePrompt + lang.autoResetPrompt}`,
+      });
+    }
+
+    // 初始赋值 startDate, endState
+    that.startDate =
+      that.startDate || (value && lay.extend({}, options.dateTime)); // 有默认值才初始化 startDate
+    that.autoCalendarModel.auto && that.autoCalendarModel();
+    that.endState =
+      !options.range || !that.rangeLinked || !!(that.startDate && that.endDate); // 初始化选中范围状态
+
+    return that;
+  }
+
+  /**
+   * 渲染备注
+   * @param {JQuery} tdElem td 元素
+   * @param {[number, number, number]} YMD 年月日
+   * @param {object | string} markers 备注信息
+   */
+  markRender(tdElem, YMD, markers) {
+    var markText;
+
+    if (typeof markers === 'object') {
+      Object.entries(markers || {}).forEach(function ([key, title]) {
+        var keys = key.split('-');
+        if (
+          (keys[0] == YMD[0] || keys[0] == 0) && // 每年的每月
+          (keys[1] == YMD[1] || keys[1] == 0) && // 每月的每日
+          keys[2] == YMD[2]
+        ) {
+          // 特定日
+          markText = title || YMD[2];
+        }
+      });
+    } else if (typeof markers === 'string') {
+      markText = markers || YMD[2];
+    }
+
+    markText &&
+      tdElem
+        .find('div')
+        .html('<span class="lay-datepicker-day-mark">' + markText + '</span>');
+  }
+
+  /**
+   * 公历重要日期与自定义备注
+   * @param {JQuery} td
+   * @param {[number, number, number]} YMD
+   * @returns Class
+   */
+  mark(td, YMD) {
+    var that = this;
+    var options = that.options;
+
+    var render = function (markers) {
+      that.markRender(td, YMD, markers);
+    };
+
+    // chineseFestivals 仅简体中文生效
+    if (options.showFestival) {
+      if (options.lang === zhCN) {
+        render(that.markerOfChineseFestivals);
+      }
+    }
+
+    if (typeof options.mark === 'function') {
+      options.mark({ year: YMD[0], month: YMD[1], date: YMD[2] }, render);
+    } else if (typeof options.mark === 'object') {
+      render(options.mark);
+    }
+
+    return that;
+  }
+
+  /**
+   * 渲染法定节假日或补假上班标记
+   * @param {JQuery} tdElem td 元素
+   * @param {[number, number, number]} YMD 年月日
+   * @param {[Array<string>, Array<string>] | string} markers 标记信息
+   */
+  holidaysRender(tdElem, YMD, markers) {
+    var type = ['holidays', 'workdays'];
+
+    var isEquals = function (ymdStr1, ymdStr2) {
+      var ymd1 = ymdStr1.split('-');
+      var ymd2 = ymdStr2.split('-');
+
+      ymd1.forEach(function (v, i) {
+        ymd1[i] = parseInt(v, 10);
+      });
+      ymd2.forEach(function (v, i) {
+        ymd2[i] = parseInt(v, 10);
+      });
+
+      return ymd1.join('-') === ymd2.join('-');
+    };
+
+    var insertHtml = function (el, type, text) {
+      el.find('div').html(
+        [
+          '<span',
+          ' class="lay-datepicker-day-holidays"',
+          ' type="' + type + '"',
+          '>',
+          text,
+          '</span>',
+        ].join(''),
+      );
+    };
+
+    if (lay.type(markers) === 'array') {
+      markers.forEach(function (item, idx) {
+        (item || []).forEach(function (dayStr) {
+          if (isEquals(dayStr, tdElem.attr('lay-ymd'))) {
+            insertHtml(tdElem, type[idx], YMD[2]);
+          }
+        });
+      });
+    } else if (typeof markers === 'string') {
+      if (type.indexOf(markers) !== -1) {
+        insertHtml(tdElem, markers, YMD[2]);
+      }
+    }
+  }
+
+  /**
+   * 标注法定节假日或补假上班
+   * @param {JQuery} td
+   * @param {[number, number, number]} YMD
+   * @returns Class
+   */
+  holidays(td, YMD) {
+    var that = this;
+    var options = that.options;
+
+    var render = function (markers) {
+      that.holidaysRender(td, YMD, markers);
+    };
+
+    if (typeof options.holidays === 'function') {
+      options.holidays({ year: YMD[0], month: YMD[1], date: YMD[2] }, render);
+    } else if (lay.type(options.holidays) === 'array') {
+      render(options.holidays);
+    }
+
+    return that;
+  }
+
+  /**
+   * 自定义单元格
+   * @param {HTMLElement|Array<HTMLElement>} el - 单元格元素
+   * @param {{year:number, month:number, date:number}} dateObj - 当前单元格对应的日期信息
+   * @param {'year' | 'month' | 'date'} panelMode - 面板模式
+   * @returns
+   */
+  cellRender(el, dateObj, panelMode) {
+    var that = this;
+    var options = that.options;
+
+    if (typeof options.cellRender === 'function') {
+      var render = function (content) {
+        if (typeof content === 'string') {
+          $(el).html(content);
+        } else if (typeof content === 'object') {
+          $(el).html('').append($(content)[0]);
+        }
+      };
+      options.cellRender(dateObj, render, { originElem: el, type: panelMode });
+    }
+
+    return that;
+  }
+
+  /**
+   * 给定年份的开始日期
+   * @param {Date} date
+   */
+  startOfYear(date) {
+    var newDate = new Date(date);
+    newDate.setFullYear(newDate.getFullYear(), 0, 1);
+    newDate.setHours(0, 0, 0, 0);
+    return newDate;
+  }
+
+  /**
+   * 给定年份的结束日期
+   * @param {Date} date
+   */
+  endOfYear(date) {
+    var newDate = new Date(date);
+    var year = newDate.getFullYear();
+    newDate.setFullYear(year + 1, 0, 0);
+    newDate.setHours(23, 59, 59, 999);
+    return newDate;
+  }
+
+  /**
+   * 给定月份的开始日期
+   * @param {Date} date
+   */
+  startOfMonth(date) {
+    var newDate = new Date(date);
+    newDate.setDate(1);
+    newDate.setHours(0, 0, 0, 0);
+    return newDate;
+  }
+
+  /**
+   * 给定月份的结束日期
+   * @param {Date} date
+   */
+  endOfMonth(date) {
+    var newDate = new Date(date);
+    var month = newDate.getMonth();
+    newDate.setFullYear(newDate.getFullYear(), month + 1, 0);
+    newDate.setHours(23, 59, 59, 999);
+    return newDate;
+  }
+
+  /**
+   * 将指定的天数添加到给定日期
+   * @param {Date} date 要更改的日期
+   * @param {number} amount 天数
+   */
+  addDays(date, amount) {
+    var newDate = new Date(date);
+    if (!amount) return newDate;
+    newDate.setDate(newDate.getDate() + amount);
+    return newDate;
+  }
+
+  /**
+   * 不可选取的年或月。年或月中的所有日期都禁用时，才判定为不可选取。
+   * @param {Date} date 要检测的年或月
+   * @param {'year' | 'month'} type 面板类型
+   * @param {'start' | 'end'} position 面板位置
+   */
+  isDisabledYearOrMonth(date, type, position) {
+    var that = this;
+    var options = that.options;
+    var millisecondsInDay = 24 * 60 * 60 * 1000;
+
+    var startDay =
+      type === 'year' ? that.startOfYear(date) : that.startOfMonth(date);
+    var endDay = type === 'year' ? that.endOfYear(date) : that.endOfMonth(date);
+    var numOfDays =
+      Math.floor((endDay.getTime() - startDay.getTime()) / millisecondsInDay) +
+      1;
+    var disabledCount = 0;
+
+    for (var i = 0; i < numOfDays; i++) {
+      var day = that.addDays(startDay, i);
+      if (options.disabledDate.call(options, day, position)) {
+        disabledCount++;
+      }
+    }
+
+    return disabledCount === numOfDays;
+  }
+
+  /**
+   * @typedef limitOptions
+   * @prop {JQuery} [elem] - 检测的元素, 例如面板中年月日时分秒元素，「现在」，「确认」 按钮等
+   * @prop {number} [index] - 元素集合中，当前检测元素的索引，years:0,month:0,date:0-41,hms:0
+   * @prop {['hours', 'minutes', 'seconds'] | ['hours', 'minutes'] | ['hours']} [time] - 是否比较时分秒
+   * @prop {'year'|'month'|string} [type] - 面板类型?
+   * @prop {0 | 1} [rangeType] - 面板索引, 0 表示 start, 1 表示 end
+   * @prop {Partial<{year:number,month: number,date:number,hours:number,minutes:number,seconds:number}>} [date] - 检测的日期时间对象
+   * @prop {'date' | 'time' | 'datetime'} disabledType - 禁用类型，按钮应使用 datetime
+   */
+  /**
+   * 不可选取的日期
+   * @param {number} date 当前检测的日期的时间戳
+   * @param {limitOptions} opts
+   * @returns {boolean}
+   */
+  isDisabledDate(date, opts) {
+    opts = opts || {};
+
+    var that = this;
+    var options = that.options;
+    var position = options.range
+      ? opts.rangeType === 0
+        ? 'start'
+        : 'end'
+      : 'start';
+
+    if (!options.disabledDate) return false;
+    if (options.type === 'time') return false;
+    if (!(opts.disabledType === 'date' || opts.disabledType === 'datetime'))
+      return false;
+
+    // 不需要时分秒
+    var normalizedDate = new Date(date);
+    normalizedDate.setHours(0, 0, 0, 0);
+
+    return opts.type === 'year' || opts.type === 'month'
+      ? that.isDisabledYearOrMonth(normalizedDate, opts.type, position)
+      : options.disabledDate.call(options, normalizedDate, position);
+  }
+
+  /**
+   * 不可选取的时间
+   * @param {number} date 当前检测的日期的时间戳
+   * @param {limitOptions} opts
+   * @returns {boolean}
+   */
+  isDisabledTime(date, opts) {
+    opts = opts || {};
+
+    var that = this;
+    var options = that.options;
+    var position = options.range
+      ? opts.rangeType === 0
+        ? 'start'
+        : 'end'
+      : 'start';
+
+    if (!options.disabledTime) return false;
+    if (!(options.type === 'time' || options.type === 'datetime')) return false;
+    if (!(opts.disabledType === 'time' || opts.disabledType === 'datetime'))
+      return false;
+
+    var isDisabledItem = function (compareVal, rangeFn, rangeFnParam) {
+      return function () {
+        return (
+          (
+            (typeof rangeFn === 'function' &&
+              rangeFn.apply(options, rangeFnParam)) ||
+            []
+          ).indexOf(compareVal) !== -1
+        );
+      };
+    };
+
+    var dateObj = that.systemDate(new Date(date));
+    var disabledTime =
+      options.disabledTime.call(options, that.newDate(dateObj), position) || {};
+
+    // 面板中的时分秒 HTML 元素需要分别检测是否禁用
+    // 按钮检测任意一项是否禁用即可
+    return opts.disabledType === 'datetime'
+      ? isDisabledItem(dateObj.hours, disabledTime.hours)() ||
+          isDisabledItem(dateObj.minutes, disabledTime.minutes, [
+            dateObj.hours,
+          ])() ||
+          isDisabledItem(dateObj.seconds, disabledTime.seconds, [
+            dateObj.hours,
+            dateObj.minutes,
+          ])()
+      : [
+          isDisabledItem(dateObj.hours, disabledTime.hours),
+          isDisabledItem(dateObj.minutes, disabledTime.minutes, [
+            dateObj.hours,
+          ]),
+          isDisabledItem(dateObj.seconds, disabledTime.seconds, [
+            dateObj.hours,
+            dateObj.minutes,
+          ]),
+        ][opts.time.length - 1]();
+  }
+
+  /**
+   * 不可选取的日期时间
+   * @param {number} timestamp 当前检测的日期的时间戳
+   * @param {limitOptions} opts
+   * @returns
+   */
+  isDisabledDateTime(timestamp, opts) {
+    opts = opts || {};
+
+    var that = this;
+    // var options = that.options;
+
+    return (
+      that.isDisabledDate(timestamp, opts) ||
+      that.isDisabledTime(timestamp, opts)
+    );
+  }
+
+  /**
+   * 无效日期范围的标记
+   * @param {limitOptions} opts
+   *
+   */
+  limit(opts) {
+    opts = opts || {};
+
+    var that = this;
+    var options = that.options;
+    var timestamp = {};
+    var dateTime =
+      opts.index > (opts.time ? 0 : 41) ? that.endDate : options.dateTime;
+    var isOut;
+
+    Object.entries({
+      now: lay.extend({}, dateTime, opts.date || {}),
+      min: options.min,
+      max: options.max,
+    }).forEach(function ([key, item]) {
+      timestamp[key] = that
+        .newDate(
+          lay.extend(
+            {
+              year: item.year,
+              month: opts.type === 'year' ? 0 : item.month, // 年份的时候只比较年
+              date:
+                opts.type === 'year' || opts.type === 'month' ? 1 : item.date, // 年月只比较年月不与最大最小比日期
+            },
+            (function () {
+              var hms = {};
+              (opts.time || []).forEach(function (keys) {
+                hms[keys] = item[keys];
+              });
+              return hms;
+            })(),
+          ),
+        )
+        .getTime(); // time：是否比较时分秒
+    });
+
+    isOut =
+      timestamp.now < timestamp.min ||
+      timestamp.now > timestamp.max ||
+      that.isDisabledDateTime(timestamp.now, opts);
+
+    opts.elem &&
+      opts.elem[isOut ? 'addClass' : 'removeClass'](CONST.CLASS_IS_DISABLED);
+
+    return isOut;
+  }
+
+  // 当前日期对象
+  thisDateTime(index) {
+    var that = this,
+      options = that.options;
+    return index ? that.endDate : options.dateTime;
+  }
+
+  // 日历表
+  calendar(value, index, type) {
+    index = index ? 1 : 0;
+    var that = this;
+    const Constructor = that.constructor;
+    var options = that.options,
+      dateTime = value || that.thisDateTime(index),
+      thisDate = new Date(),
+      startWeek,
+      prevMaxDate,
+      thisMaxDate,
+      lang = that.i18nMessages,
+      isAlone = options.type !== 'date' && options.type !== 'datetime',
+      tds = $(that.tableElems[index]).find('td'),
+      elemYM = $(that.headerElems[index][2]).find('span');
+
+    // 年份是否在有效范围内
+    if (dateTime.year < CONST.LIMIT_YEAR[0]) {
+      dateTime.year = CONST.LIMIT_YEAR[0];
+      that.hint({ content: lang.invalidDatePrompt });
+    }
+    if (dateTime.year > CONST.LIMIT_YEAR[1]) {
+      dateTime.year = CONST.LIMIT_YEAR[1];
+      that.hint({ content: lang.invalidDatePrompt });
+    }
+
+    // 记录初始值
+    if (!that.firstDate) {
+      that.firstDate = lay.extend({}, dateTime);
+    }
+
+    // 计算当前月第一天的星期
+    thisDate.setFullYear(dateTime.year, dateTime.month, 1);
+    startWeek = (thisDate.getDay() + (7 - options.weekStart)) % 7;
+
+    prevMaxDate = Constructor.getDaysInMonth(
+      dateTime.month || 12,
+      dateTime.year,
+    ); // 计算上个月的最后一天
+    thisMaxDate = Constructor.getDaysInMonth(dateTime.month + 1, dateTime.year); // 计算当前月的最后一天
+
+    // 赋值日
+    tds.each(function (index_, item) {
+      var YMD = [dateTime.year, dateTime.month],
+        st;
+      item = $(item);
+      item.removeAttr('class');
+      if (index_ < startWeek) {
+        st = prevMaxDate - startWeek + index_;
+        item.addClass('lay-datepicker-day-prev');
+        YMD = that.getAsYM(dateTime.year, dateTime.month, 'sub');
+      } else if (index_ >= startWeek && index_ < thisMaxDate + startWeek) {
+        st = index_ - startWeek;
+        if (!that.rangeLinked) {
+          st + 1 === dateTime.date && item.addClass(CONST.CLASS_IS_SELECTED);
+        }
+      } else {
+        st = index_ - thisMaxDate - startWeek;
+        item.addClass('lay-datepicker-day-next');
+        YMD = that.getAsYM(dateTime.year, dateTime.month);
+      }
+      YMD[1]++;
+      YMD[2] = st + 1;
+      item.attr('lay-ymd', YMD.join('-')).html('<div>' + YMD[2] + '</div>');
+      that
+        .mark(item, YMD)
+        .holidays(item, YMD)
+        .limit({
+          elem: item,
+          date: {
+            year: YMD[0],
+            month: YMD[1] - 1,
+            date: YMD[2],
+          },
+          index: index_,
+          rangeType: index,
+          disabledType: 'date', // 日面板，检测当前日期是否禁用
+        });
+      that.cellRender(
+        item,
+        { year: YMD[0], month: YMD[1], date: YMD[2] },
+        'date',
+      );
+    });
+
+    // 同步头部年月
+    $(elemYM[0]).attr('lay-ym', dateTime.year + '-' + (dateTime.month + 1));
+    $(elemYM[1]).attr('lay-ym', dateTime.year + '-' + (dateTime.month + 1));
+    if (!that.panelYM) that.panelYM = {};
+    that.panelYM[index] = { year: dateTime.year, month: dateTime.month };
+
+    var normalizedYearStr = addSpaceBetweenChars(
+      dateTime.year + lang.literal.year,
+    );
+    var normalizedMonthStr = addSpaceBetweenChars(lang.months[dateTime.month]);
+    if (!lang.monthBeforeYear) {
+      $(elemYM[0]).attr('lay-type', 'year').html(normalizedYearStr);
+      $(elemYM[1]).attr('lay-type', 'month').html(normalizedMonthStr);
+    } else {
+      $(elemYM[0]).attr('lay-type', 'month').html(normalizedMonthStr);
+      $(elemYM[1]).attr('lay-type', 'year').html(normalizedYearStr);
+    }
+
+    // 初始默认选择器
+    if (isAlone) {
+      // 年、月等独立选择器
+      if (options.range) {
+        if (value || type !== 'init') {
+          // 判断是否需要显示年月时间列表
+          that.listYM = [
+            [
+              (that.startDate || options.dateTime).year,
+              (that.startDate || options.dateTime).month + 1,
+            ],
+            [that.endDate.year, that.endDate.month + 1],
+          ];
+          that.list(options.type, 0).list(options.type, 1);
+
+          // 同步按钮可点状态
+          options.type === 'time'
+            ? that.setBtnStatus(
+                true,
+                lay.extend({}, that.systemDate(), that.startTime),
+                lay.extend({}, that.systemDate(), that.endTime),
+              )
+            : that.setBtnStatus(true);
+        }
+      } else {
+        that.listYM = [[dateTime.year, dateTime.month + 1]];
+        that.list(options.type, 0);
+      }
+    }
+
+    // 初始赋值双日历
+    if (options.range && type === 'init') {
+      // 执行渲染第二个日历
+      if (that.rangeLinked) {
+        var EYM = that.getAsYM(
+          dateTime.year,
+          dateTime.month,
+          index ? 'sub' : null,
+        );
+        that.calendar(
+          lay.extend({}, dateTime, {
+            year: EYM[0],
+            month: EYM[1],
+          }),
+          1 - index,
+        ); // 渲染另外一个
+      } else {
+        that.calendar(null, 1 - index);
+      }
+    }
+
+    // 通过检测当前有效日期，来设定底部按钮状态
+    if (!options.range) {
+      var timeParams = ['hours', 'minutes', 'seconds'];
+
+      // 现在按钮
+      that.limit({
+        elem: $(that.footer).find(`.${CONST.ELEM_NOW}`),
+        date: that.systemDate(
+          /^(datetime|time)$/.test(options.type) ? new Date() : null,
+        ),
+        index: 0,
+        time: timeParams,
+        disabledType: 'datetime', // 按钮，检测日期和时间
+      });
+      // 确认按钮
+      that.limit({
+        elem: $(that.footer).find(`.${CONST.ELEM_CONFIRM}`),
+        index: 0,
+        time: timeParams,
+        disabledType: 'datetime', // 按钮，检测日期和时间
+      });
+    }
+
+    // 同步按钮可点状态
+    that.setBtnStatus();
+
+    // 重置快捷栏选中状态
+    $(that.shortcut)
+      .find('li.' + CONST.CLASS_IS_SELECTED)
+      .removeClass(CONST.CLASS_IS_SELECTED);
+
+    // 标记选择范围
+    if (options.range && !isAlone && type !== 'init') that.stampRange();
+
+    return that;
+  }
+
+  // 生成年月时分秒列表
+  list(type, index) {
+    var that = this,
+      options = that.options,
+      dateTime = that.rangeLinked
+        ? options.dateTime
+        : [options.dateTime, that.endDate][index],
+      lang = that.i18nMessages,
+      isAlone =
+        options.range && options.type !== 'date' && options.type !== 'datetime', // 独立范围选择器
+      ul = lay.elem('ul', {
+        class:
+          CONST.ELEM_LIST +
+          ' ' +
+          {
+            year: 'lay-datepicker-year-list',
+            month: 'lay-datepicker-month-list',
+            time: 'lay-datepicker-time-list',
+          }[type],
+      }),
+      headerElem = that.headerElems[index],
+      elemYM = $(headerElem[2]).find('span'),
+      contentElem = that.contentElems[index || 0],
+      haveList = $(contentElem).find('.' + CONST.ELEM_LIST)[0],
+      isMonthBeforeYear = lang.monthBeforeYear,
+      text = lang.literal.year,
+      listYM = that.listYM[index] || {},
+      hms = ['hours', 'minutes', 'seconds'],
+      startEnd = ['startTime', 'endTime'][index];
+
+    if (listYM[0] < 1) listYM[0] = 1;
+
+    // 生成年列表
+    if (type === 'year') {
+      var yearNum,
+        startY = (yearNum = listYM[0] - 7);
+      if (startY < 1) startY = yearNum = 1;
+      for (var yearIndex = 0; yearIndex < 15; yearIndex++) {
+        var li = lay.elem('li', {
+            'lay-ym': yearNum,
+          }),
+          ymd = {
+            year: yearNum,
+            month: 0,
+            date: 1,
+          };
+
+        yearNum == listYM[0] && $(li).addClass(CONST.CLASS_IS_SELECTED);
+        li.innerHTML = yearNum + text;
+        ul.appendChild(li);
+
+        /*
+        if(yearNum < that.firstDate.year){
+          ymd.month = options.min.month;
+          ymd.date = options.min.date;
+        } else if(yearNum >= that.firstDate.year){
+          ymd.month = options.max.month;
+          ymd.date = options.max.date;
+        }
+        */
+
+        that.limit({
+          elem: $(li),
+          date: ymd,
+          index: index,
+          type: type,
+          rangeType: index,
+          disabledType: 'date', // 年面板，检测当前年份中的所有日期是否禁用
+        });
+        that.cellRender(li, { year: yearNum, month: 1, date: 1 }, 'year');
+        yearNum++;
+      }
+
+      $(elemYM[!isMonthBeforeYear ? 0 : 1])
+        .attr('lay-ym', yearNum - 8 + '-' + listYM[1])
+        .html(startY + text + ' - ' + (yearNum - 1 + text));
+    }
+
+    // 生成月列表
+    else if (type === 'month') {
+      for (var i = 0; i < 12; i++) {
+        const li = lay.elem('li', {
+          'lay-ym': i,
+        });
+        const ymd = {
+          year: listYM[0],
+          month: i,
+          date: 1,
+        };
+
+        i + 1 == listYM[1] && $(li).addClass(CONST.CLASS_IS_SELECTED);
+        li.innerHTML = lang.months[i];
+        ul.appendChild(li);
+
+        /*
+        if(listYM[0] < that.firstDate.year){
+          ymd.date = options.min.date;
+        } else if(listYM[0] >= that.firstDate.year){
+          ymd.date = options.max.date;
+        }
+        */
+
+        that.limit({
+          elem: $(li),
+          date: ymd,
+          index: index,
+          type: type,
+          rangeType: index,
+          disabledType: 'date', // 月面板，检测当前月份中的所有日期是否禁用
+        });
+        that.cellRender(
+          li,
+          { year: listYM[0], month: i + 1, date: 1 },
+          'month',
+        );
+      }
+
+      $(elemYM[!isMonthBeforeYear ? 0 : 1])
+        .attr('lay-ym', listYM[0] + '-' + listYM[1])
+        .html(listYM[0] + text);
+    }
+
+    // 生成时间列表
+    else if (type === 'time') {
+      // 检测时分秒状态是否在有效日期时间范围内
+      var setTimeStatus = function () {
+        $(ul)
+          .find('ol')
+          .each(function (i, ol) {
+            $(ol)
+              .find('li')
+              .each(function (ii, li) {
+                that.limit({
+                  elem: $(li),
+                  date: [
+                    {
+                      hours: ii,
+                    },
+                    {
+                      hours: that[startEnd].hours,
+                      minutes: ii,
+                    },
+                    {
+                      hours: that[startEnd].hours,
+                      minutes: that[startEnd].minutes,
+                      seconds: ii,
+                    },
+                  ][i],
+                  index: index,
+                  rangeType: index,
+                  disabledType: 'time', // 时间面板，分别检测时分秒列表是否禁用
+                  time: [
+                    ['hours'],
+                    ['hours', 'minutes'],
+                    ['hours', 'minutes', 'seconds'],
+                  ][i],
+                });
+              });
+          });
+        if (!options.range) {
+          that.limit({
+            elem: $(that.footer).find(`.${CONST.ELEM_CONFIRM}`),
+            date: that[startEnd],
+            index: 0,
+            time: ['hours', 'minutes', 'seconds'],
+            disabledType: 'datetime', // 确认按钮，检测时分秒列表任意一项是否禁用
+          });
+        }
+      };
+
+      var setTimeListVisibility = function () {
+        var showHour = options.format.indexOf('H') !== -1;
+        var showMinute = options.format.indexOf('m') !== -1;
+        var showSecond = options.format.indexOf('s') !== -1;
+        var liElem = ul.children;
+        var hideCount = 0;
+
+        [showHour, showMinute, showSecond].forEach(function (isShow, i) {
+          if (!isShow) {
+            liElem[i].className += ' lay-hide';
+            hideCount++;
+          }
+        });
+        ul.className += ' lay-datepicker-time-list-hide-' + hideCount;
+      };
+
+      // 初始化时间对象
+      if (options.range) {
+        if (!that[startEnd]) {
+          that[startEnd] = startEnd === 'startTime' ? dateTime : that.endDate;
+        }
+      } else {
+        that[startEnd] = dateTime;
+      }
+
+      // 生成时分秒
+      [24, 60, 60].forEach(function (item, i) {
+        var li = lay.elem('li'),
+          childUL = ['<p>' + lang.time[i] + '</p><ol>'];
+        for (var ii = 0; ii < item; ii++) {
+          childUL.push(
+            '<li' +
+              (that[startEnd][hms[i]] === ii
+                ? ' class="' + CONST.CLASS_IS_SELECTED + '"'
+                : '') +
+              '>' +
+              lay.digit(ii, 2) +
+              '</li>',
+          );
+        }
+        li.innerHTML = childUL.join('') + '</ol>';
+        ul.appendChild(li);
+      });
+      setTimeStatus();
+      setTimeListVisibility();
+    }
+
+    // 插入容器
+    if (haveList) contentElem.removeChild(haveList);
+    contentElem.appendChild(ul);
+
+    // 年月面板 - 选择事件
+    if (type === 'year' || type === 'month') {
+      // 显示切换箭头
+      $(that.mainListElems[index]).addClass('lay-datepicker-ym-show');
+
+      // 选中
+      $(ul)
+        .find('li')
+        .on('click', function () {
+          var ym = $(this).attr('lay-ym') | 0;
+          if ($(this).hasClass(CONST.CLASS_IS_DISABLED)) return;
+          if (that.rangeLinked) {
+            lay.extend(dateTime, {
+              year: type === 'year' ? ym : listYM[0],
+              month: type === 'year' ? listYM[1] - 1 : ym,
+            });
+          } else {
+            dateTime[type] = ym;
+          }
+
+          // 当为年选择器或者年月选择器
+          var isYearOrMonth = ['year', 'month'].indexOf(options.type) !== -1;
+          var isChangeMonth =
+            type === 'year' &&
+            ['date', 'datetime'].indexOf(options.type) !== -1;
+          if (isYearOrMonth || isChangeMonth) {
+            $(ul)
+              .find('.' + CONST.CLASS_IS_SELECTED)
+              .removeClass(CONST.CLASS_IS_SELECTED);
+            $(this).addClass(CONST.CLASS_IS_SELECTED);
+
+            // 如果为年月选择器，点击了年列表，则切换到月选择器
+            if (
+              (options.type === 'month' && type === 'year') ||
+              isChangeMonth
+            ) {
+              that.listYM[index][0] = ym;
+              isAlone && ((index ? that.endDate : dateTime).year = ym);
+              that.list('month', index);
+            }
+          } else {
+            that.checkDate('limit').calendar(dateTime, index, 'init'); // 重新渲染一下两个面板
+            that.closeList();
+          }
+
+          if (!options.range) {
+            that.limit({
+              type: type,
+              elem: $(that.footer).find(`.${CONST.ELEM_CONFIRM}`),
+              date: dateTime,
+              disabledType: 'datetime', // 按钮，检测日期和时间
+            });
+          }
+
+          that.setBtnStatus(); // 同步按钮可点状态
+
+          // 若为月选择器，只有当选择月份时才自动关闭；
+          // 若为年选择器，选择年份即自动关闭
+          // 且在范围未开启时
+          if (!options.range && options.autoConfirm) {
+            if (
+              (options.type === 'month' && type === 'month') ||
+              (options.type === 'year' && type === 'year')
+            ) {
+              that.setValue(that.parse()).done();
+              that.close();
+            }
+          }
+
+          that.autoCalendarModel.auto && !that.rangeLinked
+            ? that.choose($(contentElem).find('td.lay-this'), index)
+            : that.endState && that.done('onChange');
+          $(that.footer)
+            .find('.' + CONST.ELEM_TIME_BTN)
+            .removeClass(CONST.CLASS_IS_DISABLED);
+        });
+    } else {
+      // 时间选择面板 - 选择事件
+      var span = lay.elem('span', {
+          class: CONST.ELEM_TIME_TEXT,
+        }),
+        // 滚动条定位
+        scroll = function () {
+          $(ul)
+            .find('ol')
+            .each(function (i) {
+              var ol = this,
+                li = $(ol).find('li');
+              ol.scrollTop = 30 * (that[startEnd][hms[i]] - 2);
+              if (ol.scrollTop <= 0) {
+                li.each(function (ii) {
+                  if (!$(this).hasClass(CONST.CLASS_IS_DISABLED)) {
+                    ol.scrollTop = 30 * (ii - 2);
+                    return false;
+                  }
+                });
+              }
+            });
+        },
+        haveSpan = $(headerElem[2]).find('.' + CONST.ELEM_TIME_TEXT);
+
+      scroll();
+      span.innerHTML = options.range
+        ? [lang.startTime, lang.endTime][index]
+        : lang.selectTime;
+      $(that.mainListElems[index]).addClass('lay-datepicker-time-show');
+
+      if (haveSpan[0]) {
+        haveSpan.remove();
+      }
+      headerElem[2].appendChild(span);
+
+      var olElem = $(ul).find('ol');
+      olElem.each(function (i) {
+        var ol = this;
+        // 选择时分秒
+        $(ol)
+          .find('li')
+          .on('click', function () {
+            var value = this.innerHTML | 0;
+            if ($(this).hasClass(CONST.CLASS_IS_DISABLED)) return;
+
+            if (options.range) {
+              that[startEnd][hms[i]] = value;
+            } else {
+              dateTime[hms[i]] = value;
+            }
+            $(ol)
+              .find('.' + CONST.CLASS_IS_SELECTED)
+              .removeClass(CONST.CLASS_IS_SELECTED);
+            $(this).addClass(CONST.CLASS_IS_SELECTED);
+
+            setTimeStatus();
+            scroll();
+            (that.endDate ||
+              options.type === 'time' ||
+              options.type === 'datetime') &&
+              that.done('onChange');
+
+            // 同步按钮可点状态
+            that.setBtnStatus();
+          });
+      });
+
+      if (lay.device().mobile) {
+        olElem.css({
+          overflowY: 'auto',
+          touchAction: 'pan-y',
+        });
+      }
+    }
+
+    return that;
+  }
+
+  // 记录列表切换后的年月
+  listYM = [];
+
+  // 关闭列表
+  closeList() {
+    var that = this;
+    // var options = that.options;
+
+    that.contentElems.forEach(function (item, index) {
+      $(item)
+        .find('.' + CONST.ELEM_LIST)
+        .remove();
+      $(that.mainListElems[index]).removeClass(
+        'lay-datepicker-ym-show lay-datepicker-time-show',
+      );
+    });
+    that.$mainElem.find('.' + CONST.ELEM_TIME_TEXT).remove();
+  }
+
+  // 检测结束日期是否超出开始日期
+  setBtnStatus(tips, start, end) {
+    var that = this,
+      options = that.options,
+      lang = that.i18nMessages,
+      isOut,
+      elemBtn = $(that.footer).find(`.${CONST.ELEM_CONFIRM}`),
+      timeParams =
+        options.type === 'datetime' || options.type === 'time'
+          ? ['hours', 'minutes', 'seconds']
+          : undefined;
+    if (options.range) {
+      start = start || (that.rangeLinked ? that.startDate : options.dateTime);
+      end = end || that.endDate;
+      isOut =
+        !that.endState ||
+        that.newDate(start).getTime() > that.newDate(end).getTime();
+
+      // 如果不在有效日期内，直接禁用按钮，否则比较开始和结束日期
+      that.limit({
+        date: start,
+        disabledType: 'datetime', // 按钮，检测日期和时间
+        time: timeParams,
+        rangeType: 0,
+      }) ||
+      that.limit({
+        date: end,
+        disabledType: 'datetime', // 按钮，检测日期和时间
+        time: timeParams,
+        rangeType: 1,
+      })
+        ? elemBtn.addClass(CONST.CLASS_IS_DISABLED)
+        : elemBtn[isOut ? 'addClass' : 'removeClass'](CONST.CLASS_IS_DISABLED);
+
+      // 是否异常提示
+      if (tips && isOut) {
+        that.hint({ content: lang.rangeOrderPrompt });
+      }
+    }
+  }
+
+  // 转义为规定格式的日期字符
+  parse(state, date) {
+    var that = this;
+    const Constructor = this.constructor;
+    var options = that.options;
+    var startDate = that.rangeLinked ? that.startDate : options.dateTime;
+    var dateTime =
+      date ||
+      (state == 'end'
+        ? lay.extend({}, that.endDate, that.endTime)
+        : options.range
+          ? lay.extend({}, startDate || options.dateTime, that.startTime)
+          : options.dateTime);
+    var format = Constructor.formatDateTime(dateTime, that.format, 1);
+
+    // 返回日期范围字符
+    if (options.range && state === undefined) {
+      return format + ' ' + that.rangeStr + ' ' + that.parse('end');
+    }
+
+    return format;
+  }
+
+  // 创建指定日期时间对象
+  newDate(dateTime) {
+    dateTime = dateTime || {};
+    return new Date(
+      dateTime.year || 1,
+      dateTime.month || 0,
+      dateTime.date || 1,
+      dateTime.hours || 0,
+      dateTime.minutes || 0,
+      dateTime.seconds || 0,
+    );
+  }
+
+  // 获得指定日期时间对象的毫秒数
+  getDateTime(obj) {
+    return this.newDate(obj).getTime();
+  }
+
+  /**
+   * 格式化输入框显示值
+   * @param {HTMLInputElement} elem HTML input 元素
+   * @param {(value: string) => string} displayValueCallback
+   */
+  formatToDisplay(elem, displayValueCallback) {
+    var that = this;
+    var props = Object.getOwnPropertyDescriptor(
+      HTMLInputElement.prototype,
+      'value',
+    );
+
+    Object.defineProperty(
+      elem,
+      'value',
+      lay.extend({}, props, {
+        get: function () {
+          return this.getAttribute('lay-datepicker');
+        },
+        set: function (value) {
+          props.set.call(this, displayValueCallback.call(that, value));
+          this.setAttribute('lay-datepicker', value);
+        },
+      }),
+    );
+  }
+
+  // 赋值
+  setValue(value) {
+    var that = this;
+    var options = that.options;
+    var $elem = options.$elem;
+
+    // 内嵌模式不作默认赋值
+    if (options.inline) return that;
+
+    value = value || '';
+
+    // 绑定的元素是否为 input
+    if (that.isInput($elem[0])) {
+      $elem.val(value);
+    } else {
+      // 如果 range 传入了开始和结束的 input 对象，则分别对其赋值
+      var rangeElem = that.rangeElem;
+      if (rangeElem) {
+        if (lay.type(value) !== 'array') {
+          value = value.split(' ' + that.rangeStr + ' ');
+        }
+        rangeElem[0].val(value[0] || '');
+        rangeElem[1].val(value[1] || '');
+      } else {
+        if ($elem.find('*').length === 0) {
+          var displayValue =
+            typeof options.formatToDisplay === 'function'
+              ? options.formatToDisplay(value)
+              : value;
+          $elem.html(displayValue);
+        }
+        $elem.attr('lay-datepicker', value);
+      }
+    }
+
+    return that;
+  }
+
+  // 预览
+  preview() {
+    var that = this,
+      options = that.options;
+
+    if (!options.showPreview) return;
+
+    var elemPreview = that.$mainElem.find('.' + CONST.ELEM_PREVIEW),
+      value = options.range
+        ? (that.rangeLinked ? that.endState : that.endDate)
+          ? that.parse()
+          : ''
+        : that.parse();
+
+    // 显示预览
+    elemPreview.html(value);
+
+    // 预览颜色渐变
+    var oldValue = elemPreview.html();
+    if (oldValue) {
+      var color =
+        lay.type(options.variant) === 'array'
+          ? options.variant[0]
+          : options.variant;
+      elemPreview.css({
+        color: /^#/.test(String(color)) ? color : '#16b777',
+      });
+
+      setTimeout(function () {
+        elemPreview.css({
+          color: '#777',
+        });
+      }, 300);
+    }
+  }
+
+  // 附加的渲染处理，在 ready 和 onChange 的时候调用
+  renderAdditional() {
+    var that = this;
+    var options = that.options;
+
+    // 处理全面板
+    if (options.fullPanel) {
+      that.list('time', 0);
+    }
+  }
+
+  // 标记范围内的日期
+  stampRange() {
+    var that = this,
+      options = that.options,
+      startTime = that.rangeLinked ? that.startDate : options.dateTime,
+      endTime,
+      tds = that.$mainElem.find('td');
+
+    if (options.range && !that.endState)
+      $(that.footer)
+        .find(`.${CONST.ELEM_CONFIRM}`)
+        .addClass(CONST.CLASS_IS_DISABLED);
+    // if(!that.endState) return;
+
+    startTime =
+      startTime &&
+      that
+        .newDate({
+          year: startTime.year,
+          month: startTime.month,
+          date: startTime.date,
+        })
+        .getTime();
+
+    endTime =
+      that.endState &&
+      that.endDate &&
+      that
+        .newDate({
+          year: that.endDate.year,
+          month: that.endDate.month,
+          date: that.endDate.date,
+        })
+        .getTime();
+
+    // if(startTime > endTime) return that.hint({ content: TIPS_OUT });
+
+    tds.each(function (i, item) {
+      var ymd = $(item).attr('lay-ymd').split('-');
+      var thisTime = that
+        .newDate({
+          year: ymd[0],
+          month: ymd[1] - 1,
+          date: ymd[2],
+        })
+        .getTime();
+
+      // 标记当天
+      if (options.rangeLinked && !that.startDate) {
+        if (thisTime === that.newDate(that.systemDate()).getTime()) {
+          $(item).addClass(
+            $(item).hasClass(CONST.ELEM_PREV) ||
+              $(item).hasClass(CONST.ELEM_NEXT)
+              ? ''
+              : CONST.ELEM_DAY_NOW,
+          );
+        }
+      }
+
+      /*
+       * 标注区间
+       */
+
+      $(item).removeClass(CONST.ELEM_SELECTED + ' ' + CONST.CLASS_IS_SELECTED);
+
+      if (thisTime === startTime || thisTime === endTime) {
+        (that.rangeLinked ||
+          (!that.rangeLinked &&
+            (i < 42 ? thisTime === startTime : thisTime === endTime))) &&
+          $(item).addClass(
+            $(item).hasClass(CONST.ELEM_PREV) ||
+              $(item).hasClass(CONST.ELEM_NEXT)
+              ? CONST.ELEM_SELECTED
+              : CONST.CLASS_IS_SELECTED,
+          );
+      }
+      if (thisTime > startTime && thisTime < endTime) {
+        $(item).addClass(CONST.ELEM_SELECTED);
+      }
+    });
+  }
+
+  // 执行 done/onChange 回调
+  done(type, param) {
+    var that = this;
+    var options = that.options;
+    var start = lay.extend(
+      {},
+      lay.extend(
+        that.rangeLinked ? that.startDate : options.dateTime,
+        that.startTime,
+      ),
+    );
+    var end = lay.extend({}, lay.extend(that.endDate, that.endTime));
+
+    [start, end].forEach(function (item) {
+      if (!('month' in item)) return;
+      lay.extend(item, {
+        month: item.month + 1,
+      });
+    });
+
+    that.preview();
+
+    const params = {
+      value: that.parse(),
+      start: start,
+      end: end,
+      ...param,
+    };
+
+    type === 'onChange' && that.renderAdditional();
+    options[type || 'done']?.(params);
+
+    return that;
+  }
+
+  /**
+   * 判断日期是否存在面板，用于处理日期范围选择的一些边缘情况
+   * @param {object} datetime 日期时间对象
+   * @param {number} index - 面板索引
+   * @returns
+   */
+  checkPanelDate(datetime, index) {
+    var that = this;
+    var options = that.options;
+    // 年月范围选择不需要处理
+    if (options.type !== 'date' && options.type !== 'datetime') return;
+
+    var startPanel = index === 0;
+    var month = datetime.month + 1; // 点击的日期所在月份
+    var panelMonth = that.panelYM[index].month + 1; // 当前面板头部月份
+
+    // 边缘日期的处理
+    var firstRenderIndex = that.endState
+      ? // 二次点击（一般为结束日期）任意一侧面板时：
+        // 1. 左侧面板中，点击的月份属于下一个月时，应渲染右侧面板而不是左侧面板;
+        // 2. 左侧面板中，点击的月份属于上一个月时，应将两个面板都重新渲染(等效点击 prevMonth);
+        // 3. 右侧面板同理。
+        (function () {
+          return (startPanel && month > panelMonth) ||
+            (!startPanel && month < panelMonth)
+            ? 1 - index
+            : index;
+        })()
+      : // 初次点击（一般为开始日期）任意一侧面板时：
+        // 1. 让该面板自行切换，以保持日期的「选中状态」在该侧；
+        // 2. 另一侧面板则根据点击的面板进行响应式切换，以保持左右面板始终为连续月份。
+        index;
+
+    // 为简化实现，只要点击的月份不等于当前面板顶部显示的月份时，就重新渲染两侧面板
+    return {
+      needFullRender: month !== panelMonth,
+      index: firstRenderIndex,
+    };
+  }
+
+  // 选择日期
+  choose(td, index) {
+    if (td.hasClass(CONST.CLASS_IS_DISABLED)) return;
+
+    var that = this,
+      options = that.options,
+      panelIndex = index; // 记录点击的是哪一个面板的
+
+    if (that.rangeLinked) {
+      if (that.endState || !that.startDate) {
+        // 重新选择或者第一次选择
+        index = 0;
+        that.endState = false;
+      } else {
+        index = 1;
+        that.endState = true;
+      }
+    }
+
+    var dateTime = that.thisDateTime(index),
+      // tds = that.$mainElem.find('td'),
+      YMD = td.attr('lay-ymd').split('-');
+
+    YMD = {
+      year: YMD[0] | 0,
+      month: (YMD[1] | 0) - 1,
+      date: YMD[2] | 0,
+    };
+
+    lay.extend(dateTime, YMD); // 同步 dateTime
+
+    // 范围选择
+    if (options.range) {
+      // 补充时分秒
+      ['startTime', 'endTime'].forEach(function (item, i) {
+        that[item] = that[item] || {
+          hours: i ? 23 : 0,
+          minutes: i ? 59 : 0,
+          seconds: i ? 59 : 0,
+        };
+        if (index === i) {
+          // 判断选择之后的是否在范围内，超出则需要调整时分秒
+          if (
+            that.getDateTime(lay.extend({}, dateTime, that[item])) <
+            that.getDateTime(options.min)
+          ) {
+            that[item] = {
+              hours: options.min.hours,
+              minutes: options.min.minutes,
+              seconds: options.min.seconds,
+            };
+            lay.extend(dateTime, that[item]);
+          } else if (
+            that.getDateTime(lay.extend({}, dateTime, that[item])) >
+            that.getDateTime(options.max)
+          ) {
+            that[item] = {
+              hours: options.max.hours,
+              minutes: options.max.minutes,
+              seconds: options.max.seconds,
+            };
+            lay.extend(dateTime, that[item]);
+          }
+        }
+      });
+      if (!index) {
+        that.startDate = lay.extend({}, dateTime); // 同步 startDate
+      }
+      // 校验另外一个日期是否在有效的范围内
+      // 此处为范围选择的日期面板点击选中处理，所以 disabledType 为 date
+      if (
+        that.endState &&
+        !that.limit({
+          date: that.rangeLinked
+            ? that.startDate
+            : that.thisDateTime(1 - index),
+          disabledType: 'date',
+        })
+      ) {
+        // 根据选择之后判断是否需要切换模式
+        var isChange;
+        if (that.endState && that.autoCalendarModel.auto) {
+          isChange = that.autoCalendarModel();
+        }
+        // 判断是否反选
+        var needSwapDate =
+          (isChange || (that.rangeLinked && that.endState)) &&
+          that.newDate(that.startDate) > that.newDate(that.endDate);
+        if (needSwapDate) {
+          var isSameDate =
+            that.startDate.year === that.endDate.year &&
+            that.startDate.month === that.endDate.month &&
+            that.startDate.date === that.endDate.date;
+          var startDate;
+          // 如果是同一天并且出现了反选证明是时分秒出现开始时间大于结束时间的现象
+          if (isSameDate) {
+            startDate = that.startTime;
+            that.startTime = that.endTime;
+            that.endTime = startDate;
+          }
+          // 当出现反向选择时（即「后点击」的日期比「先点击」的日期小），重新提取区间
+          startDate = that.startDate;
+          that.startDate = lay.extend({}, that.endDate, that.startTime);
+          options.dateTime = lay.extend({}, that.startDate);
+          that.endDate = lay.extend({}, startDate, that.endTime);
+        }
+        isChange && (options.dateTime = lay.extend({}, that.startDate));
+      }
+      if (that.rangeLinked) {
+        // 处理日期范围选择的一些边缘情况
+        var checkState = that.checkPanelDate(dateTime, panelIndex);
+        var dateTimeTemp = lay.extend({}, dateTime);
+        var renderMode =
+          isChange || (checkState && checkState.needFullRender) ? 'init' : null;
+        var panelIdx = checkState ? checkState.index : panelIndex;
+        that.calendar(dateTimeTemp, panelIdx, renderMode);
+      } else {
+        that.calendar(null, index, isChange ? 'init' : null);
+      }
+      that.endState && that.done('onChange');
+    } else if (options.inline) {
+      // 直接嵌套的选中
+      that.calendar().done().done('onChange'); // 同时执行 done 和 onChange 回调
+    } else if (options.type === 'date') {
+      if (options.autoConfirm) {
+        that.setValue(that.parse()).done();
+        that.close();
+      } else {
+        that.calendar().done('onChange');
+      }
+    } else if (options.type === 'datetime') {
+      that.calendar().done('onChange');
+    }
+  }
+
+  // 底部按钮
+  tool(btn, type) {
+    var that = this,
+      options = that.options,
+      lang = that.i18nMessages,
+      dateTime = options.dateTime,
+      active = {
+        // 选择时间
+        datetime: function () {
+          if ($(btn).hasClass(CONST.CLASS_IS_DISABLED)) return;
+          that.list('time', 0);
+          options.range && that.list('time', 1);
+          $(btn).attr('lay-type', 'date').html(that.i18nMessages.selectDate);
+        },
+
+        // 选择日期
+        date: function () {
+          that.closeList();
+          $(btn)
+            .attr('lay-type', 'datetime')
+            .html(that.i18nMessages.selectTime);
+        },
+
+        // 清空、重置
+        clear: function () {
+          options.inline &&
+            (lay.extend(dateTime, that.firstDate), that.calendar());
+          options.range &&
+            (delete options.dateTime,
+            delete that.endDate,
+            delete that.startTime,
+            delete that.endTime);
+          that.setValue('');
+          that.done(null, 'onClear').done({
+            value: '',
+            start: {},
+            end: {},
+          });
+          that.close();
+        },
+
+        // 现在
+        now: function () {
+          var thisDate = new Date();
+
+          // 当前系统时间未在 min/max 范围内，则不可点击
+          if ($(btn).hasClass(CONST.CLASS_IS_DISABLED)) {
+            return that.hint({
+              content: lang.tools.now + ', ' + lang.invalidDatePrompt,
+            });
+          }
+
+          lay.extend(dateTime, that.systemDate(), {
+            hours: thisDate.getHours(),
+            minutes: thisDate.getMinutes(),
+            seconds: thisDate.getSeconds(),
+          });
+
+          that.setValue(that.parse());
+          options.inline && that.calendar();
+          that.done(null, 'onNow').done(null, 'onChange');
+          that.close();
+        },
+
+        // 确定
+        confirm: function () {
+          if (options.range) {
+            if ($(btn).hasClass(CONST.CLASS_IS_DISABLED)) {
+              var isTimeout =
+                options.type === 'time'
+                  ? that.startTime &&
+                    that.endTime &&
+                    that.newDate(that.startTime) > that.newDate(that.endTime)
+                  : that.startDate &&
+                    that.endDate &&
+                    that.newDate(
+                      lay.extend({}, that.startDate, that.startTime || {}),
+                    ) >
+                      that.newDate(
+                        lay.extend({}, that.endDate, that.endTime || {}),
+                      );
+
+              return that.hint({
+                content: isTimeout
+                  ? lang.rangeOrderPrompt
+                  : lang.invalidDatePrompt,
+              });
+            }
+          } else {
+            if ($(btn).hasClass(CONST.CLASS_IS_DISABLED))
+              return that.hint({ content: lang.invalidDatePrompt });
+          }
+
+          that.setValue(that.parse());
+          that.done(null, 'onConfirm').done();
+          that.close();
+        },
+      };
+    active[type] && active[type]();
+  }
+
+  // 统一切换处理
+  change(index) {
+    var that = this,
+      options = that.options,
+      dateTime = that.thisDateTime(index),
+      isAlone =
+        options.range && (options.type === 'year' || options.type === 'month'),
+      contentElem = that.contentElems[index || 0],
+      listYM = that.listYM[index],
+      addSubYear = function (type) {
+        var isYear = $(contentElem).find('.lay-datepicker-year-list')[0],
+          isMonth = $(contentElem).find('.lay-datepicker-month-list')[0];
+
+        // 切换年列表
+        if (isYear) {
+          listYM[0] = type ? listYM[0] - 15 : listYM[0] + 15;
+          that.list('year', index);
+        }
+
+        if (isMonth) {
+          // 切换月面板中的年
+          type ? listYM[0]-- : listYM[0]++;
+          that.list('month', index);
+        }
+
+        if (isYear || isMonth) {
+          lay.extend(dateTime, {
+            year: listYM[0],
+          });
+          if (isAlone) dateTime.year = listYM[0];
+          options.range || that.done(null, 'onChange');
+          options.range ||
+            that.limit({
+              elem: $(that.footer).find(`.${CONST.ELEM_CONFIRM}`),
+              date: {
+                year: listYM[0],
+                month: isYear ? 0 : listYM[1] - 1,
+              },
+              disabledType: 'datetime', // 按钮，检测日期和时间
+            });
+        }
+
+        that.setBtnStatus();
+        return isYear || isMonth;
+      };
+
+    return {
+      prevYear: function () {
+        if (addSubYear('sub')) return;
+        if (that.rangeLinked) {
+          options.dateTime.year--;
+          that.checkDate('limit').calendar(null, null, 'init');
+        } else {
+          dateTime.year--;
+          that.checkDate('limit').calendar(null, index);
+          // 面板自动切换的模式下重新判定是否发生模式转换等细节处理
+          that.autoCalendarModel.auto
+            ? that.choose($(contentElem).find('td.lay-this'), index)
+            : that.done(null, 'onChange');
+        }
+      },
+      prevMonth: function () {
+        var YM;
+        // rangeLinked 模式非实时选择日期，不需要同步 options.dateTime，应根据面板显示日期切换
+        if (that.rangeLinked) {
+          var panelYM = that.panelYM[0];
+          var dateTimeTemp;
+          YM = that.getAsYM(panelYM.year, panelYM.month, 'sub');
+          dateTimeTemp = lay.extend({}, options.dateTime, that.panelYM[0], {
+            year: YM[0],
+            month: YM[1],
+          });
+          that.checkDate('limit').calendar(dateTimeTemp, null, 'init');
+        } else {
+          YM = that.getAsYM(dateTime.year, dateTime.month, 'sub');
+          lay.extend(dateTime, {
+            year: YM[0],
+            month: YM[1],
+          });
+
+          that.checkDate('limit').calendar(null, null, 'init');
+          that.autoCalendarModel.auto
+            ? that.choose($(contentElem).find('td.lay-this'), index)
+            : that.done(null, 'onChange');
+        }
+      },
+      nextMonth: function () {
+        var YM;
+
+        if (that.rangeLinked) {
+          var panelYM = that.panelYM[0];
+          YM = that.getAsYM(panelYM.year, panelYM.month);
+          var dateTimeTemp = lay.extend({}, options.dateTime, that.panelYM[0], {
+            year: YM[0],
+            month: YM[1],
+          });
+          that.checkDate('limit').calendar(dateTimeTemp, null, 'init');
+        } else {
+          YM = that.getAsYM(dateTime.year, dateTime.month);
+
+          lay.extend(dateTime, {
+            year: YM[0],
+            month: YM[1],
+          });
+
+          that.checkDate('limit').calendar(null, null, 'init');
+          that.autoCalendarModel.auto
+            ? that.choose($(contentElem).find('td.lay-this'), index)
+            : that.done(null, 'onChange');
+        }
+      },
+      nextYear: function () {
+        if (addSubYear()) return;
+        if (that.rangeLinked) {
+          options.dateTime.year++;
+          that.checkDate('limit').calendar(null, 0, 'init');
+        } else {
+          dateTime.year++;
+          that.checkDate('limit').calendar(null, index);
+          that.autoCalendarModel.auto
+            ? that.choose($(contentElem).find('td.lay-this'), index)
+            : that.done(null, 'onChange');
+        }
+      },
+    };
+  }
+
+  // 日期切换事件
+  changeEvent() {
+    var that = this;
+    // var options = that.options;
+
+    // 阻止日期面板内的点击/按下事件冒泡
+    that.$mainElem.on('click mousedown', function (e) {
+      e.stopPropagation();
+    });
+
+    // 年月切换
+    that.headerElems.forEach(function (header, i) {
+      // 上一年
+      $(header[0]).on('click', function () {
+        that.change(i).prevYear();
+      });
+
+      // 上一月
+      $(header[1]).on('click', function () {
+        that.change(i).prevMonth();
+      });
+
+      // 选择年月
+      $(header[2])
+        .find('span')
+        .on('click', function () {
+          var othis = $(this),
+            layYM = othis.attr('lay-ym'),
+            layType = othis.attr('lay-type');
+
+          if (!layYM) return;
+
+          layYM = layYM.split('-');
+
+          that.listYM[i] = [layYM[0] | 0, layYM[1] | 0];
+          that.list(layType, i);
+          $(that.footer)
+            .find('.' + CONST.ELEM_TIME_BTN)
+            .addClass(CONST.CLASS_IS_DISABLED);
+        });
+
+      // 下一月
+      $(header[3]).on('click', function () {
+        that.change(i).nextMonth();
+      });
+
+      // 下一年
+      $(header[4]).on('click', function () {
+        that.change(i).nextYear();
+      });
+    });
+
+    // 点击日期
+    that.tableElems.forEach(function (table, i) {
+      var tds = $(table).find('td');
+      tds.on('click', function () {
+        that.choose($(this), i);
+      });
+    });
+
+    // 点击底部按钮
+    $(that.footer)
+      .find('span')
+      .on('click', function () {
+        var type = $(this).attr('lay-type');
+        that.tool(this, type);
+      });
+  }
+
+  // 是否输入框
+  isInput(elem = this.options.$elem?.[0]) {
+    if (!elem) return false;
+    return (
+      /input|textarea/.test(elem.tagName.toLocaleLowerCase()) ||
+      /INPUT|TEXTAREA/.test(elem.tagName)
+    );
+  }
+
+  // 绑定的元素事件处理
+  events() {
+    /* var that = this;
+    var options = that.options;
+
+    if (!options.$elem[0] || options.$elem[0].eventHandler) return;
+
+    var showEvent = function () {
+      // 已经打开的面板避免重新渲染
+      if (datePicker.thisId === options.id) return;
+      that.render();
+    };
+
+    // 绑定呼出控件事件
+    options.$elem.on(options.trigger, showEvent);
+    options.$elem[0].eventHandler = true;
+    options.$eventElem.on(options.trigger, showEvent);
+
+    // 元素解绑
+    that.unbind = function () {
+      that.close();
+      options.$elem.off(options.trigger, showEvent);
+      options.$elem.removeAttr('lay-key');
+      options.$elem.removeAttr(MOD_ID);
+      options.$elem[0].eventHandler = false;
+      options.$eventElem.off(options.trigger, showEvent);
+      options.$eventElem.removeAttr('lay-key');
+      delete thisModule.that[options.id];
+    }; */
+  }
+}
+
+const CONST = DatePicker.CONST;
+
+// 简体中文语言码
+const zhCN = 'zh-CN';
+
+// 年份在前的语言
+const YearBeforeMonthLocale = [
+  'eu-ES',
+  'ja-JP',
+  'km-KH',
+  'ko-KR',
+  'pt-BR',
+  'si-LK',
+  'ms-MY',
+  'ug-CN',
+  'zh-CN',
+  'zh-HK',
+  'zh-TW',
+];
+
+/**
+ * 在字符之间添加空格
+ * @param {string} str
+ * @returns {string}
+ */
+const addSpaceBetweenChars = (str) => {
+  if (typeof str !== 'string' || str.length <= 1) {
+    return str;
+  }
+
+  const isDigit = (char) => {
+    var code = char.charCodeAt(0);
+    return code >= 48 && code <= 57; // '0' 到 '9' 的 ASCII 码范围
+  };
+
+  var result = '';
+  for (var i = 0; i < str.length - 1; i++) {
+    var char = str[i];
+    var nextChar = str[i + 1];
+    result += char;
+
+    // 判断当前字符和下一个字符的类型
+    var isCharDigit = isDigit(char);
+    var isNextCharDigit = isDigit(nextChar);
+
+    // 在数字和非数字（非空格）之间添加空格
+    if (
+      (isCharDigit && !isNextCharDigit && nextChar !== ' ') || // 数字 → 非数字（非空格）
+      (char !== ' ' && !isCharDigit && isNextCharDigit) // 非空格非数字 → 数字
+    ) {
+      result += ' ';
+    }
+  }
+  result += str[str.length - 1]; // 添加最后一个字符
+  return result;
+};
+
+// 默认日期时间格式
+const defaultFormat = {
+  year: 'yyyy',
+  month: 'yyyy-MM',
+  date: 'yyyy-MM-dd',
+  time: 'HH:mm:ss',
+  datetime: 'yyyy-MM-dd HH:mm:ss',
+};
+
+// 日期格式占位符正则模式
+const DATE_FORMAT_REGEX = 'yyyy|y|MM|M|dd|d|HH|H|mm|m|ss|s';
+
+// 将日期格式字符拆分为数组
+const splitDateFormat = (format) => {
+  return (format || '').match(new RegExp(DATE_FORMAT_REGEX + '|.', 'g')) || [];
+};
+
+export { DatePicker as datePicker };
