+++ src/core/i18n.js
@@ -47,7 +47,7 @@ var zhCN = {
     },
     verifyErrorPromptTitle: '提示',
   },
-  laydate: {
+  datePicker: {
     months: [
       '1月',
       '2月',
