+++ src/css/index.css
@@ -34,5 +34,5 @@
 @import './modules/tree.css';
 @import './modules/anim.css';
 @import './modules/code.css';
-@import './modules/laydate.css';
+@import './modules/datepicker.css';
 @import './modules/layer.css';
