+++ tests/visual/popup.html
@@ -204,7 +204,7 @@
         popup.render({
           elem: '#demo-backdrop',
           content:
-            'Popup 是附着弹出、选择器等类型组件（如：dropdown、laydate、colorPicker）的基础组件。',
+            'Popup 是附着弹出、选择器等类型组件（如：dropdown、datePicker、colorPicker）的基础组件。',
           style: 'width: 350px;',
           backdrop: 0.5,
           placement: 'bottom-start',
